Now the Admnis Options and the User Settings are refactorized in Compound Components Design Pattern and in the same way will be supported With Component Composition.
Basically this new way to refactoring the code shared contexts in the components and use customize hooks to administrate the whole states, effects and the handling functions in the same context.
This is usefull because now we don´t need two forms to add or update a new Product, Collection, Provider, etc...

The same form use the context  to choose if there is an information to update, or, if there is no information or data to update, then the will create a new   product, collection or what the admin requires to manipulate

Use this Design Pattern makes more easy to read the code since it is divided into different files  and in the case that a dev needs to add a new functionality.
Now he can make the code upragade so much easier to do, this because now is so much more scalable than without using this pattern.
The context and the props of a the component now can be re adapted to a new necessitie without breaking the code.